+++ example/clock/main.cpp
@@ -555,7 +555,7 @@ public:
 
     }
 
-    void onMouseButton(int btn,int action,int mode,int x,int y) override
+    void onMouseButton(int btn,int action,int mode,double x,double y) override
     {
         RenderDemoRotate::onMouseButton(btn,action,mode,x,y);
         if(btn == GLFW_MOUSE_BUTTON_2 && action == GLFW_PRESS)
