+++ example/music_player/main.cpp
@@ -467,7 +467,7 @@ public:
 
     }
 
-    void onMouseButton(int btn,int action,int mode,int x,int y) override
+    void onMouseButton(int btn,int action,int mode,double x,double y) override
     {
         RenderDemoRotate::onMouseButton(btn,action,mode,x,y);
         if(action == GLFW_PRESS)
