+++ include/RenderDemo.h
@@ -33,7 +33,7 @@ public:
     virtual int init();
     virtual void run();
     virtual void onWindowResize(int w,int h);
-    virtual void onMouseButton(int,int,int,int,int);
+    virtual void onMouseButton(int,int,int, double, double);
     virtual void onMouseMove(double,double);
 
     WINDOW_TYPE get_window();
@@ -65,7 +65,7 @@ protected:
     static void unregOnMouseButtonListener(RenderDemo* rd);
 
     static void MouseButtonCallBack(WINDOW_TYPE,int,int,int);
-    static void MouseButtonCallBackEx(WINDOW_TYPE, int, int, int, int, int);
+    static void MouseButtonCallBackEx(WINDOW_TYPE, int, int, int, double, double);
 //-----------------------------------------------------------
     static std::vector<RenderDemo*> MouseMoveListeners;
